requirements.txt
Flask==3.0.0
flask-cors==4.0.0
RPi.GPIO==0.7.1
adafruit-circuitpython-ads1x15==2.2.21
adafruit-circuitpython-dht==4.0.3
requests==2.31.0
